Move my intcode computer code out into a library and make existing code use it
As of the end of Day 9, we apparently have a complete implementation of an Intcode virtual machine. It seems likely we're going to get a few more intcode-related problems, so this seemed like a good point to make that code reusable.

I now have an intcode subdirectory containing the source code for my intcode vm implementation. I've modified each of my existing intcode-related solutions to use that instead of their existing copy 'n' pasted implementations.