Add a python implementation of the intcode vm
I wanted to try writing the intcode VM as a python generator. It worked out quite similar to my "tiny" c++ implementation, the main difference being that we can write it as a single self-contained function rather than needing to make it a struct. The instruction pointer and relative address offset can be local variables in the function, all we have to pass in is the program.

The VM yields whenever it requires input. You supply input to it using the generator's built-in `send()` method. It also yields whenever it writes output; the output is the value it yields.

This ends up being a little awkward to use, unfortunately. For problems which receive input (i.e. most of them) you can't simply use a for loop. You need to have the generator object in order to send input to it. It gets worse with problems like day 11, when the program writes multiple output values in sequence. It might be nicer to buffer up output values until the next input request, or something.