Make the python intcode vm buffer up output until a read
Modified the python version of my intcode VM so that it only ever yields when it needs input. The value we yield at that point is a list containing all output since the last yield. This means we no longer need to yield a status code as well, which simplifies things for client code.

On the day 11 problems, we no longer have to make multiple calls the iterator's `next()` and `send()` methods. Inside our calling loop we now only make a single call to `send`.

I think this version is easier to use, but it's still a bit ugly: you still have to call `g.next()` once to make the VM run up to the first input instruction, for example. And you still can't use it with a for loop, because calling `send()` advances the iterator.