Optimise day16b
When applying the pattern, we were doing two forward iterations over the array: one to calculate a sum and one to store values, subtracting from the sum as it went. This has been replaced with a single reverse iteration which writes values and adds to the sum as it goes. This ends up reducing the overall runtime by about 60ms, from 170 to 110 ms.

Now storing only the part of the (repeated) input array that we need for the calculation. This reduces the overall runtime by a further 30 ms to approximately 80 ms.