Refactor tiny-intcode.h to take a queue of input values
Instead of taking a single input value, the vh::IntcodeComputer::run() method now takes a std::deque of them and will keep running until all available inputs are exhausted or the program halts.

I've added overloads of run() which take 0 and 1 inputs, respectively. These just populate a std::deque and call through to the main run() method.

I also added an overload which supports passing a null-terminated string as input. This populates a std::deque with the characters of the string and passes it to the main run() method. This overload comes in handy for day 17.