Add code to find the subroutines for day17b
Wrote a python prototype which finds the subroutines and the main path, given an existing full path. Ported this to c++ and integrated it into my day 17 solution, so now it should work on any of the official puzzle inputs (I think).

It fails on the challenge path given here though: https://www.reddit.com/r/adventofcode/comments/ebz338/2019_day_17_part_2_pathological_pathfinding/ I'm not sure why yet, I think it's because the input path I'm generating isn't sufficiently compressible. Need to have a way of trying other paths if we can't find a solution for the first path we generate.

A release build solves the day 17b problem in 9 milliseconds.