Add an interactive intcode runner
This is a simple REPL which runs an intcode program, printing any output it produces to the terminal and passing whatever you type back to it as input.

It has support for (very basic) scripting: you can provide one or more files containing lists of inputs. We read data from each of these files one line at a time and provide the whole line (including the trailing newline) to the intcode VM as input. When the first input file is exhausted we move on to the next, until all of them are used up. We then start getting input from stdin, until the intcode program has terminated. I used this for the day 25 problem, but it seems like a generally useful feature.