Add intcode VM benchmarks
This is a suite of intcode VM benchmarks from here:

https://www.reddit.com/r/adventofcode/comments/egq9xn/2019_day_9_intcode_benchmarking_suite/

I added a simple application to run the benchmarks too, called intcode-benchmark. It loads the intcode program specified as the first command line argument and treats all subsequent command line arguments as numeric inputs. It runs the program and reports the time between then and when it halts. These times are what you can see in the results table in the README.md file for the benchmarks directory.