Improve speed of the intcode vm
These changes give a 1.51x - 1.52x speedup in all of the test cases.

Instead of calculating the addressing mode for each argument separately we now do the lookup for all args at once. We have a switch statement which returns a pointer to one of a small number of precalculated mode arrays. This is based on the observation that there are only 18 valid combinations of addressing modes for all three arguments (bearing in mind that mode 1 is not valid for the destination argument, so can never be used for argument 3). This we do a single div and mod per instruction now, instead of up to three of each.

Replaced the address resolution logic in peek(), which consisted of two ternary if expressions, with a switch. This gives a surprisingly big speedup over the previous version, due (I think) to replacing the evaluation of two boolean expressions with a single computed jump.

Casting the opcode to an int before doing the div and mod gives a small performance increase. I tried using std::div to get the remainder and quotient simultaneously, but this turned out to be a lot slower for some reason.

Oddly the profile build ("release with debug info" in cmake) is faster than a release build!