mmap a large memory region instead of using a std::vector
We now simply mmap a large amount of memory up front, without committing it, instead of using a std::vector for storage. This means the OS's memory manager automatically handles adding new pages as necessary and the pages will be zeroed when we get them. This, in turn, means we don't need any out-of-bounds checks and we never need to resize the memory. Overall it ends up being a really big performance improvement (see the README.md in intcode/benchmarks for the updated benchmark results).