Use computed gotos to speed up the intcode vm
Instead of a switch inside a while loop, I'm now using computed gotos to jump to the correct handling code for each instruction.

This reduces the amount of work for each instruction. The old code was doing three jumps per instruction: jump to the handling code for the instruction, jump to the end of the switch statement, then jump back to the start of the loop (not sure whether the compiler combined the last two into a single jump. This code does a single jump per instruction: after executing the handling code it jumps directly to the handling code for the next instruction. The overall effect is a significant performance increase, but the code has turned into the stuff of nightmares.

The dispatch table (an array containing the address of the goto label corresponding to each instruction) is initialised the first time we enter the run() method. This seems to have an overhead of about 100 microseconds (micro, not milli). To avoid including it in the benchmark we run a trivial program, consisting of only the halt instruction, at vm construction time. This forces the dispatch table to be populated before we start the timer for the benchmark.