Add copy and move constructors to the intcode vm struct
Also added copy and move assignment operators.

Beware with copy construction/assignment: we will only copy over the bytes containing the original program, not any additional bytes beyond that range which have been written to during program execution. The proper way to deal with this is probably to set the memory up so that it's shared using copy-on-write, but I haven't tried that yet.